# docs/swagger.yaml
basePath: /
definitions:
  ErrorResponse:
    properties:
      message:
        type: string
      param:
        type: string
      value:
        type: string
    type: object
...
  models.Booking:
    properties:
      bookingId:
        type: integer
      bookingStatusId:
        type: integer
      created:
        type: string
      createdBy:
        type: integer
      lastUpdated:
        type: string
    required:
    - bookingStatusId
    - created
    - createdBy
    - lastUpdated
    type: object
  models.Match:
    properties:
      bookingId:
        type: integer
      cost:
        type: number
      created:
        type: string
      isPaid:
        type: boolean
      lastUpdated:
        type: string
      matchAccessStatusId:
        type: integer
      matchDate:
        type: string
      matchId:
        type: integer
      matchStatusId:
        type: integer
    required:
    - bookingId
    - cost
    - created
    - isPaid
    - lastUpdated
    - matchAccessStatusId
    - matchDate
    - matchStatusId
    type: object
  models.MatchPlayer:
    properties:
      amountToPay:
        type: number
      matchId:
        type: integer
      matchPlayerId:
        type: integer
      paymentTypeId:
        type: integer
      playerId:
        type: integer
    required:
    - amountToPay
    - matchId
    - paymentTypeId
    - playerId
    type: object
...
  payloads.BookingRequest:
    properties:
      day:
        maxLength: 10
        minLength: 3
        type: string
      endTime:
        type: string
      matchAccessStatusId:
        type: integer
      noOfWeeks:
        type: integer
      pitchId:
        type: integer
      squadId:
        type: integer
      startTime:
        type: string
      venueId:
        type: integer
    required:
    - day
    - endTime
    - matchAccessStatusId
    - noOfWeeks
    - pitchId
    - startTime
    - venueId
    type: object
...
host: 127.0.0.1:3000
info:
  contact:
    email: nathan.mcalinden@unosquare.com
    name: Footpal Support
  description: This is swagger for Footpal
  termsOfService: http://swagger.io/terms/
  title: Footpal API
  version: "1.0"
paths:
  /bookings:
    get:
      description: Retrieve all bookings by user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Booking'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - booking
    post:
      description: Create new single or recurring booking testing
      parameters:
      - description: Request
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/payloads.BookingRequest'
      produces:
      - application/json
      responses:
        "202":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - booking
  /bookings/{bookingId}:
    delete:
      description: Cancel active or pending booking
      parameters:
      - description: Booking ID
        in: path
        name: bookingId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - booking
    get:
      description: Retrieve booking by bookingId
      parameters:
      - description: Booking ID
        in: path
        name: bookingId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Booking'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - booking
    put:
      description: Edit booking details
      parameters:
      - description: Booking ID
        in: path
        name: bookingId
        required: true
        type: integer
      - description: Request
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/payloads.BookingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Booking'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - booking
  /bookings/{bookingId}/matches:
    get:
      description: Get all matches linked to booking
      parameters:
      - description: Booking ID
        in: path
        name: bookingId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Match'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - booking
...
  /matches:
    get:
      description: Retrieve all public matches
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Match'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - match
  /matches/{matchId}:
    delete:
      description: Cancel match
      parameters:
      - description: Match ID
        in: path
        name: matchId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - match
    get:
      description: Retrieve match details by matchId
      parameters:
      - description: Match ID
        in: path
        name: matchId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Match'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - match
  /matches/{matchId}/players:
    get:
      description: Retrieve all players opted into a match
      parameters:
      - description: Match ID
        in: path
        name: matchId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MatchPlayer'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - match
  /matches/{matchId}/players/{playerId}:
    delete:
      description: As a booking/squad admin, remove player from match
      parameters:
      - description: Match ID
        in: path
        name: matchId
        required: true
        type: integer
      - description: Player ID
        in: path
        name: playerId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      tags:
      - match
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"